refactor android files
